youngforum/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-camel-case==1.0.3
drf-yasg==1.14.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pytz==2018.9
requests==2.21.0
ruamel.yaml==0.15.89
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
